requirements.txt
# Corporate Loan Assessment RAG System Requirements
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# LLM and Embeddings
openai>=1.0.0
langchain-core>=0.2.38

# Vector Store
faiss-cpu>=1.8.0.post1

# News API
newsapi-python>=0.2.6

# Utilities
tenacity>=8.4.2
aiohttp>=3.9.5
pydantic>=2.0.0

# Data processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ML and Analysis
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Text Processing
nltk>=3.8.0